requirements.txt
psycopg2-binary
openpyxl
pandas
schedule
python-dotenv
